Toggle Sidebar Project
This project involves creating a sidebar navigation for a website, allowing users to access different sections or links of the site conveniently. The sidebar is designed to be hidden initially and can be shown or hidden by clicking on a toggle button. A close button within the sidebar allows users to easily close the sidebar if they no longer need it.