Modal project
This project is a simple example of how to create a modal dialog using HTML, CSS, and JavaScript. Modals are commonly used to provide additional information or functionality in a non-intrusive way on websites. The JavaScript code allows for the modal to be shown and hidden when appropriate user actions are taken.